requirements.txt
Django==1.3.1
PyYAML==3.10
Werkzeug==0.8.2
distribute==0.6.24
django-extensions==0.7.1
django-mptt==0.5.2
docutils==0.8.1
epydoc==3.0.1
pysqlite==2.6.3
wsgiref==0.1.2
